+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 **Added**
 
+- #187 Allow to introduce the Age instead of Date of Birth for Patient
 
 **Changed**